requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.1
pandas==2.1.4
python-dotenv==1.0.0
sqlparse==0.4.4
tqdm==4.66.1
loguru==0.7.2
click==8.1.7
pyyaml==6.0.1
